+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Remove Drush commands
 - Add sub-theme in /var/starter_kits/italiagov
 - Add ddev start script in /var/bin/start-dev-test.sh for test in docker container
+- Update README.md
+- Add comments in var/bin/start-dev-test.sh
 
 # Summary 8.x-0.8
 - Removing compatibility with Core 9.x